% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_maturities.R
\encoding{UTF-8}
\name{calculate_maturities}
\alias{calculate_maturities}
\title{Beregn forfall fra gitt startdato}
\usage{
calculate_maturities(
  df,
  start_date,
  settlementdate,
  maturitydate,
  transaction_volume,
  grouping_variables,
  type = "volume",
  freq = "d"
)
}
\arguments{
\item{df}{dataframe}

\item{start_date}{start-dato man vil beregne forfall fra}

\item{settlementdate}{navn på kolonne med oppgjørsdato}

\item{maturitydate}{navn på kolonne med forfallsdato}

\item{transaction_volume}{navn på kolonne med transaksjonsvolum}

\item{grouping_variables}{vektor med variabelnavn man vil gruppere etter}

\item{type}{char med type serie man vil velge. "volume" for forfall i volum per dag,
"cummulative volume" for kummulativt volum per dag,
"percent of outstanding" for å regne om volum per dag til prosent av utestående
volum. "cummulative percent of outstanding" for å regne ut kummulativ prosentvise
forfall av uteståend volum}
}
\description{
Beregn forfall fra gitt startdato
}
\examples{
\dontrun{m  = calculate_maturities(df,
                                   "2020-01-01",
                                   "SettlementDate",
                                   "MaturityDate",
                                   "PrincipalAmount",
                                   "Maturity") }
}
