% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_outstanding.R
\name{calculate_outstanding}
\alias{calculate_outstanding}
\title{Beregn utestående volum}
\usage{
calculate_outstanding(
  df,
  settlementdate,
  maturitydate,
  transactionvolume,
  grouping_variables,
  freq = "d"
)
}
\arguments{
\item{df}{dataframe med transaksjonsdata}

\item{settlementdate}{navn på kolonne som inneholder oppgjørsdato}

\item{maturitydate}{navn på kolonne som inneholder forfallsdato}

\item{transactionvolume}{navn på kolonne som inneholder transaksjonsvolum}

\item{grouping_variables}{vektor av variabler man skal sortere etter. Må være ett eller flere kolonnenavn i df.}

\item{freq}{frekvens man vil ha sluttresultatet på. Tar en av verdiene i "d", "b", "w", "m", "q", "y" for hhv.
daglig, business day, ukentlig, månedlig, kvartalsvis og årlig frekvens}
}
\description{
Beregn utestående volum
}
\examples{
\dontrun{dtcc_data = dtcc_data \%>\%  calculate_outstanding_amounts(settlementdate = "settlementdate",
                                       maturitydate = "maturitydate",
                                       transactionvolume = "principalamount",
                                       grouping_variables = c("producttype", "name"),
                                       freq = "w") }
}
