% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expose_dots.R
\encoding{UTF-8}
\name{expose_dots}
\alias{expose_dots}
\title{Utvid datasett}
\usage{
expose_dots(df, ...)
}
\arguments{
\item{df}{dataframe}

\item{...}{navn på variabler}
}
\description{
Utvid datasett
}
\examples{
\dontrun{df = df \%>\% expose_dots(var1, var2, var3)}
}
