% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expose_vec.R
\encoding{UTF-8}
\name{expose_vec}
\alias{expose_vec}
\title{Fyll ut datasett}
\usage{
expose_vec(df, vars, datecol)
}
\arguments{
\item{df}{Dataframe med}

\item{vars}{vektor med variabelnavn}

\item{datecol}{navn på datokolonne}
}
\description{
\code{expose_vec} fyller ut datoaksen for et datasett med ufullstendig datoakse.
Verdier på datoer uten data i opprinnelig datasett blir satt til NA. NB: datoaksen blir
satt til daglig datoakse. Det kan gi store datasett
}
\examples{
\dontrun{df = df \%>\% expose_vec(c("var1", "var2"), datecol = "date")}
}
