% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importer_bloomberg_data.R
\encoding{UTF-8}
\name{importer_bloomberg_data}
\alias{importer_bloomberg_data}
\title{Hent data fra bloomberg}
\usage{
importer_bloomberg_data(
  start_date,
  fields = c("PX_LAST"),
  tickers,
  include_non_trading_days = T,
  fill_na = T
)
}
\arguments{
\item{start_date}{start-dato for uttrekk}

\item{fields}{hvilke felter man vil ha. Default er bid, ask og last}

\item{tickers}{hvilke serier man skal hente ut.}

\item{include_non_trading_days}{Skal man inkludere helger og andre dager det ikke handles på? T eller F}

\item{fill_na}{Skal man fylle NA med forrige verdi? T eller F}
}
\description{
Hent data fra bloomberg
}
\examples{
\dontrun{df = importer_bloomberg_data(start_date = "2020-01-01",
                                      fields = c("px_last", "px_bid"),
                                      tickers = c("nok curncy", "nok3m curncy"))}


}
