% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noma_add_line.R
\encoding{UTF-8}
\name{noma_add_line}
\alias{noma_add_line}
\title{Legg til linje i plot}
\usage{
noma_add_line(p, df, xcol, ycol, p_xcol, name = ycol, color = NULL)
}
\arguments{
\item{p}{plot man skal legge til linje i. Det bør være generert med noma_tidy_plot (eller noma_ts_plot (ikke testa p.t.))}

\item{df}{dataframe med data}

\item{xcol}{navn på x-aksekollonne}

\item{ycol}{navn på y-aksekollonne}

\item{p_xcol}{navn på x-aksekollonne i p, dvs. plottet man legger til en linje i}

\item{name}{navn på linje som skal vises i legend.}

\item{color}{farge på linje}
}
\description{
Legg til linje i plot
}
\examples{
\dontrun{#' #Legge til linje fra datasett med navn bloom_data i plot ved navn p:
p = noma_add_line(p,
                  bloom_data,
                 "date",
                 "us0003m_index_px_last",
                  p_xcol =
                  "SettlementDate",
                  name = "Libor")}
}
