% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noma_agg_plot_fx.R
\name{noma_agg_plot_fx}
\alias{noma_agg_plot_fx}
\title{Plot aggregated fx swap data}
\usage{
noma_agg_plot_fx(
  df,
  counterparties = c("FOREIGN", "DOMESTIC"),
  series = "netto",
  tenors = levels(as.factor(df$tenor)),
  currencies = levels(as.factor(df$FxFrgnCcy)),
  grouping = "tenor",
  date_start = "2020-03-01",
  date_end = Sys.Date() - 1,
  yrange = NULL
)
}
\arguments{
\item{df}{Data frame med data som skal plottes}

\item{counterparties}{Velg hvilke motparter som skal plottes mot}

\item{series}{hvilken sluttserie skal vises? Tar verdiene "netto", "innlan" og "utlan"}

\item{tenors}{Hvilke løpetider skal vises? Default er alle i datasettet som brukes som input}

\item{currencies}{Hvilke valutaer skal inkluderes? Default er alle}

\item{grouping}{Hvilken variabel skal grupperes over? Default er løpetid ("tenor")}

\item{date_start}{startdato for graf}

\item{date_end}{sluttdato for graf}

\item{yrange}{egendefinert y-range (optional)}
}
\value{

}
\description{
Plot aggregated fx swap data
}
\examples{
\dontrun{
agglist <- noma_agg_plot_fx(fx, "FOREIGN", "netto", c("3m", "6m"))
agglist$p 
}
}
