% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noma_ts_plot.R
\encoding{UTF-8}
\name{noma_ts_plot}
\alias{noma_ts_plot}
\title{Plot tidsseriedata}
\usage{
noma_ts_plot(
  df,
  series_list,
  xaxis_title = "",
  yaxis_title = "",
  plot_title = "",
  legend_options = list(showlegend = T, legend_position_x = NULL, legend_position_y =
    NULL)
)
}
\arguments{
\item{df}{dataframe med tidsseriedata, dvs. en datokolonne, mange kolonner med variabler (non-tidy).}

\item{series_list}{liste med spesifikasjoner for serier som skal plottes.}

\item{xaxis_title}{navn på x-akse}

\item{yaxis_title}{navn på y-akse}

\item{plot_title}{tittel på graf}

\item{legend_options}{alternativer for legend}
}
\description{
Plot tidsseriedata
}
\examples{
\dontrun{p = noma_ts_plot(df, series_list)}
}
